Kenken Puzzle Initial Commit
Create puzzle_cell.py which is the class file for a puzzle cell. Each
puzzle cell contains the number and its notes. Utility functions like
setting notes, clearing notes, setting number are also in the file.

Create kenken_puzzle.py which is the class file for the Kenken Puzzle.
Kenken Puzzle is a square grid puzzle. Each row and column contains
number from 1 to N where N is the size of the puzzle without duplicates.
Additionally, certain cells in the puzzles will form groups/lines in
which the numbers in the cells will form up to the total of the group
depending on the operator. These are represented by lines in the code.
There is also a utility function that helps to print the puzzle.