Add solve function
Solve function that works by finding naked singles. Naked singles are
cells with only one note and hence can only have one possible number.

Add utility function, is_solved, to determine the the puzzle is solved.
Puzzle is considered solve if every cell in the puzzle contains a
number.

Notes are set when line clues are added. Line clues are mathematical
equations that apply to the cells in the line. All the cells in the line
will perform an operation on each other to get the result in the clue.
The function only works with (and assumed) the "+" operator. Possible
notes are determine as long as it is able to find a combination from all
the cells in the line to get the result.